Rename easter egg player
