+++ lib/classes/easter_egg_player.dart
