+++ lib/screens_pages/settings_page.dart
@@ -9,7 +9,7 @@ import 'package:simple_chat/utils/colorimetry_blueprints.dart';
 
 import 'package:simple_chat/utils/logger_config.dart';
 
-import '../classes/player.dart';
+import '../classes/easter_egg_player.dart';
 
 class settings extends StatefulWidget {
   const settings({super.key});
